Beans properly constructed, with various bean for classes added

+++ src/main/java/Godwin/restaurant_menu/entities/Topping.java
@@ -10,6 +10,8 @@ import lombok.ToString;
 public class Topping {
     private String name;
     private int calories;
-    private int price;
+    private double price;
+
+
 
 }
